Extract test methods to PlanUtils
